Remove :rig_api OTP app and move code into :rig
