+++ apps/rig/.gitignore
@@ -14,3 +14,6 @@ erl_crash.dump
 # secrets files as long as you replace their contents by environment
 # variables.
 /config/*.secret.exs
+
+# The Swagger JSON document is re-generated during `mix compile`:
+/priv/static/rig_api_swagger.json
