+++ apps/rig/config/rig_api/dev.exs
@@ -1,6 +1,6 @@
 use Mix.Config
 
-config :rig_api, RigApi.Endpoint,
+config :rig, RigApi.Endpoint,
   # debug_errors: true, # Uncomment to see full error descriptions in API as HTML
   env: :dev,
   check_origin: false,
