+++ apps/rig/lib/rig_api/controllers/session_blacklist_controller.ex
