+++ apps/rig/lib/rig_api/views/error_view.ex
