--- apps/rig_api/.formatter.exs
@@ -1,4 +1,0 @@
-# Used by "mix format"
-[
-  inputs: ["mix.exs", "{config,lib,test}/**/*.{ex,exs}"]
-]
