--- apps/rig_api/.gitignore
@@ -1,20 +1,0 @@
-# App artifacts
-/_build
-/db
-/deps
-/*.ez
-
-# Generated on crash by the VM
-erl_crash.dump
-
-# Files matching config/*.secret.exs pattern contain sensitive
-# data and you should not commit them into version control.
-#
-# Alternatively, you may comment the line below and commit the
-# secrets files as long as you replace their contents by environment
-# variables.
-/config/*.secret.exs
-
-# The Swagger JSON document is re-generated during `mix compile`:
-/priv/static/rig_api_swagger.json
-
