--- apps/rig_api/README.md
@@ -1,3 +1,0 @@
-# RigApi
-
-See the [developer's guide](../../guides/developer-guide.md).
