--- apps/rig_api/test/test_helper.exs
@@ -1,3 +1,0 @@
-ExUnit.start()
-# Exclude all smoke tests from running by default
-ExUnit.configure(exclude: [smoke: true])
