+++ apps/rig_tests/test/proxy/request_logger/kafka_test.exs
@@ -21,7 +21,7 @@ defmodule RigTests.Proxy.RequestLogger.KafkaTest do
   alias Rig.KafkaConfig, as: RigKafkaConfig
   alias RigKafka
 
-  @api_port Confex.fetch_env!(:rig_api, RigApi.Endpoint)[:http][:port]
+  @api_port Confex.fetch_env!(:rig, RigApi.Endpoint)[:http][:port]
   @proxy_port Confex.fetch_env!(:rig_inbound_gateway, RigInboundGatewayWeb.Endpoint)[:http][:port]
   @env [port: 55_001]
